BVTCK-125 Separate value extractors injection tests to avoid side effects
