+++ tests/src/main/resources/org/hibernate/beanvalidation/tck/tests/integration/cdi/managedobjects/validation-ManagedObjectsTest.xml
@@ -19,7 +19,4 @@
     <parameter-name-provider>org.hibernate.beanvalidation.tck.tests.integration.cdi.managedobjects.ParameterNameProviderUsingDependencyInjection</parameter-name-provider>
     <clock-provider>org.hibernate.beanvalidation.tck.tests.integration.cdi.managedobjects.ClockProviderUsingDependencyInjection</clock-provider>
 
-    <value-extractor>org.hibernate.beanvalidation.tck.tests.integration.cdi.managedobjects.MapKeyValueExtractorUsingDependencyInjection</value-extractor>
-    <value-extractor>org.hibernate.beanvalidation.tck.tests.integration.cdi.managedobjects.MapValueValueExtractorUsingDependencyInjection</value-extractor>
-
 </validation-config>
